+++ app/src/main/java/woowacourse/omok/domain/omokboard/ColumnPosition.kt
