+++ app/src/main/java/woowacourse/omok/domain/omokboard/Intersection.kt
