+++ app/src/main/java/woowacourse/omok/domain/omokboard/IntersectionState.kt
