+++ app/src/main/java/woowacourse/omok/domain/omokboard/OmokBoard.kt
