+++ app/src/main/java/woowacourse/omok/domain/omokboard/RowPosition.kt
