+++ app/src/main/java/woowacourse/omok/domain/player/PlayerStone.kt
