+++ app/src/main/java/woowacourse/omok/domain/player/StoneColor.kt
