+++ app/src/main/java/woowacourse/omok/domain/rule/OmokResult.kt
