+++ app/src/main/java/woowacourse/omok/domain/rule/OmokRule.kt
