+++ app/src/main/java/woowacourse/omok/domain/rule/judge/JudgeRule.kt
@@ -1,4 +1,4 @@
-package omok.domain.rule.winning
+package woowacourse.omok.domain.rule.judge
 
 import omok.domain.rule.OmokRule
