+++ app/src/main/java/woowacourse/omok/domain/rule/place/ExternalRule.kt
