+++ app/src/main/java/woowacourse/omok/domain/rule/place/PlaceResult.kt
